Rename logger module to "logging"
This is more accurate: the module does not itself contain a logger.
